% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acomb.R
\name{acomb}
\alias{acomb}
\title{Combine the simulated samples, used in the foreach loop}
\usage{
acomb(..., along)
}
\arguments{
\item{...}{Any number of lists, each list is either length one that only includes the array of mean matrices, or length two with both array of mean matrices and array of covariance matrices.}

\item{along}{The dimension along which to bind the arrays in the list.}
}
\value{
A list of either length one that only includes the array of mean matrices, or length two with both array of mean matrices and array of covariance matrices.
}
\description{
Combine the simulated samples, used in the foreach loop
}
\keyword{internal}
