% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commutatorTest.R
\name{commutatorTest}
\alias{commutatorTest}
\title{Test on the commutator of two matrices}
\usage{
commutatorTest(
  mat.arr,
  cn,
  eps = NULL,
  d = dim(mat.arr)[2],
  cov.arr = NULL,
  testType = c("chi", "gam")
)
}
\arguments{
\item{mat.arr}{The array of two matrices to be tested}

\item{cn}{Constant for convergence}

\item{eps}{The threshold of eigenvalues when compute general inverse of covariance matrices. Must be supplied when \code{testType = 'chi'}}

\item{d}{Size of matrices}

\item{cov.arr}{The array of covariance matrices of the two matrices, default will use identity matrices}

\item{testType}{The test methods. Either for exact chi-squared test, or approximated gamma test}
}
\value{
A list of test information.
\itemize{
\item testType Test method, \code{'chi'} or \code{'gam'}.
\item statistic Test statistic.
\item df Degrees of freedom for chi-squared distribution.
\item shape Shape parameter in gamma distribution.
\item rate Rate parameter in gamma distribution.
\item pvalue P-value.
}
}
\description{
Test on the commutator of two matrices
}
\examples{
means = generateMeans(5,2)
samples = simuSamples(means, sqrt(100), 1)
commutatorTest(samples[[1]][1,1,1,,,], sqrt(400), 400^(-1/3))
}
