% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countryCoeff}
\alias{countryCoeff}
\title{Macroeconomic VAR Model Data}
\format{
The array of coefficient matrices for country-wise macroeconomic VAR(1) models.
}
\usage{
countryCoeff
}
\description{
Estimated coefficient matrices of 8 countries' VAR(1) models.
See \insertCite{xu2021testing;textual}{eigTest}.
}
\examples{
\dontrun{
 countryCoeff
}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
