% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigTest.R
\name{eigTest}
\alias{eigTest}
\title{Test Whole Set of Eigenvectors}
\usage{
eigTest(
  A,
  cov.arr = NULL,
  cn,
  eps = NULL,
  V = JDTE(A),
  d = dim(A)[2],
  p = dim(A)[1],
  testType = c("chi", "gam"),
  param.out = FALSE
)
}
\arguments{
\item{A}{Array of matrices}

\item{cov.arr}{List of covariance matrices, default with use identity matrices}

\item{cn}{Constant for convergence}

\item{eps}{The threshold of eigenvalues when compute general inverse of covariance matrices. Must be supplied when \code{testType = 'chi'}}

\item{V}{Eigenvector to be tested}

\item{d}{Size of matrices}

\item{p}{Number of matrices}

\item{testType}{The test methods. Either for exact chi-squared test, or approximated gamma test.}

\item{param.out}{The parameters of limiting distribution should be output or not}
}
\value{
P-value or a list of test information.
\itemize{
\item testType Test method, \code{'chi'} or \code{'gam'}.
\item statistic Test statistic.
\item df Degrees of freedom for chi-squared distribution.
\item shape Shape parameter in gamma distribution.
\item rate Rate parameter in gamma distribution.
\item pvalue P-value.
}
}
\description{
Test Whole Set of Eigenvectors
}
\examples{
eigTest(countryCoeff, countryCovar, cn = sqrt(112), testType = 'gam')
}
