% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMeans.R
\name{generateMeans}
\alias{generateMeans}
\title{Generate Mean Matrices for Simulations}
\usage{
generateMeans(d, p, k = d, snr = NULL, V = NULL, v = NULL, nn = FALSE)
}
\arguments{
\item{d}{The dimension of matrices.}

\item{p}{The number of matrices.}

\item{k}{The number of common Schur components. Must be an integer within (0, \code{d}), otherwise set \code{k = d}.}

\item{snr}{The positive signal to noise variance ratio (SNR), can be a vector. Default \code{is.null(snr) = TRUE} will only generate means with non-perturbed eigenvectors.}

\item{V}{The input of eigenvector matrix with dimension \code{d}-\code{d}, needed when \code{nn = FALSE}. Default will use random sampling when \code{is.null(V) = TRUE}.}

\item{v}{The input of stationary distribution with length \code{d}, needed when \code{nn = TRUE}. Default will use Dirichlet random sampling when \code{is.null(v) = TRUE}.}

\item{nn}{Logical, whether the generated matrices should be nonnegative as transition probability matrices.}
}
\value{
An array of mean matrices: \code{p}-by-\code{q}-by-\code{d}-by-\code{d}, where \code{q} is the number of SNRs: \code{q = length(snr) + 1}.
}
\description{
Mean matrices with (randomly perturbed) common eigenvectors are generated for simulations setups.
See \insertCite{xu2021testing;textual}{eigTest}.
}
\examples{
generateMeans(5,8,3)
}
\references{
\insertAllCited{}
}
