% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hudsonDaily}
\alias{hudsonDaily}
\title{Hudson river daily discharge table.}
\format{
The table of the Hudson river daily discharge with daily quantile and categories.
}
\usage{
hudsonDaily
}
\description{
Hudson river daily discharge table.
}
\examples{
\dontrun{
 hudsonDaily
}
}
\keyword{datasets}
