% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hudsonWeekly}
\alias{hudsonWeekly}
\title{Hudson river weekly discharge table.}
\format{
The table of the Hudson river weekly discharge with weekly quantile and categories.
}
\usage{
hudsonWeekly
}
\description{
Hudson river weekly discharge table.
}
\examples{
\dontrun{
 hudsonWeekly
}
}
\keyword{datasets}
