% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.Q.R
\name{init.Q}
\alias{init.Q}
\title{Initiate orthogonal matrix \code{Q}.}
\usage{
init.Q(A)
}
\arguments{
\item{A}{Array of matrices}
}
\value{
Best initial condition for function \code{partSchur}.
}
\description{
Initiate orthogonal matrix \code{Q}.
}
\keyword{internal}
