% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert}
\alias{invert}
\title{Matrix inverse wrapper}
\usage{
invert(X)
}
\arguments{
\item{X}{The square matrix to be inverted.}
}
\value{
The matrix inverse of \code{X}.
}
\description{
Prefer \code{solve}. When singular error is encountered, use \code{ginv} instead.
}
\keyword{internal}
