% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnPartSchur.R
\name{nnPartSchur}
\alias{nnPartSchur}
\title{Joint Schur decomposition for the non-negative leading component}
\usage{
nnPartSchur(A)
}
\arguments{
\item{A}{The array of matrices with dimension \code{p}-\code{d}-\code{d}, where \code{p} is the number of matrices, \code{d} is the dimension of the matrices.}
}
\value{
The orthogonal matrix \code{Q} of dimension \code{d}-\code{d} with the first column to be the estimated nonnegative common component.
}
\description{
Joint Schur decomposition for the non-negative leading component
}
\examples{
nnPartSchur(countryCoeff)
}
\references{
\insertRef{xu2021testing}{eigTest}
}
