% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialTest.R
\name{partialTest}
\alias{partialTest}
\title{Test Subset of Common Eigenvectors}
\usage{
partialTest(
  A,
  cn,
  cov.arr = NULL,
  nn = FALSE,
  k = NULL,
  warmup = FALSE,
  Q = NULL,
  testType = c("chi", "gam"),
  eps = NULL,
  param.out = FALSE
)
}
\arguments{
\item{A}{Array of original matrices}

\item{cn}{Constant for convergence}

\item{cov.arr}{Array of original covariance matrices, default will use identity matrices}

\item{nn}{Logical whether the eigenvector elements are nonnegative}

\item{k}{Size of the block matrices to be tested}

\item{warmup}{Logical whether use \code{partSchur} for a warm-up \code{Q}}

\item{Q}{Schur components for transformation}

\item{testType}{The test methods. Either for exact chi-squared test, or approximated gamma test}

\item{eps}{The threshold of eigenvalues when compute general inverse of covariance matrices. Required when \code{testType = 'chi'} but with default \code{cn^(-2/3)} when unsupplied}

\item{param.out}{The parameters of limiting distribution should be output or not}
}
\value{
A named vector of P-value(s) when \code{param.out = FALSE}, or a named list of test information when \code{param.out = TRUE}, with name(s) to be \code{'chi'} and/or \code{'gam'}. Test information is a sub-list with elements including:
\itemize{
\item statistic Test statistic.
\item df Degrees of freedom for chi-squared distribution.
\item shape Shape parameter in gamma distribution.
\item rate Rate parameter in gamma distribution.
\item pvalue P-value.
}
}
\description{
Test whether a subset of common eigenvectors are shared by the list of random matrices.
}
\examples{
partialTest(countryCoeff, cn = sqrt(112), cov.arr = countryCovar, k = 2, testType = 'gam')
}
