% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partSchur.R
\name{partSchur}
\alias{partSchur}
\title{Partial Joint Schur Decomposition}
\usage{
partSchur(
  A,
  k,
  n = ncol(A[[1]]),
  p = length(A),
  iter = 5000,
  tol = 10^(-16),
  nonneg = FALSE
)
}
\arguments{
\item{A}{List of matrices}

\item{k}{Number of Schur components}

\item{n}{Size of matrices}

\item{p}{Number of matrices}

\item{iter}{Maximum iteration number}

\item{tol}{Tolerance error}

\item{nonneg}{Logical whether the eigenvector elements are nonnegative}
}
\value{
Orthogonal matrix \code{Q}
}
\description{
Partial Joint Schur Decomposition
}
\examples{
partSchur(countryCoeff, k = 2)
}
