% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.fun.R
\name{score.fun}
\alias{score.fun}
\title{Score function (off-diagonal squared norm or first semi-column squared norm)}
\usage{
score.fun(A, Q = NULL)
}
\arguments{
\item{A}{The array of matrices with dimension \code{p}-\code{d}-\code{d}, where \code{p} is the number of matrices, \code{d} is the dimension of the matrices.}

\item{Q}{Optional. The orthogonal matrix of dimension \code{d}-\code{d} with the first column evaluated for common. Default (\code{is.null(Q) = TRUE}) will be ignored for off-diagonal norm.}
}
\value{
Either the sum of squared off-diagonal norms of \code{Ai} or the sum of squared norms of \code{t(Q[,-1]) Ai Q[,1]}.
}
\description{
Calculate either the summed squared off-diagonal 'F' norms of \code{Ai} or the summed squared 'F' norms of \code{Bi[(j+1):n,j]} where \code{Bi = t(Q) Ai  Q} when \code{Q} is supplied.
}
\keyword{internal}
