% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuSamples.R
\name{simuSamples}
\alias{simuSamples}
\title{Generating consistent estimators for given means}
\usage{
simuSamples(mu, cn, reps, nonneg = FALSE)
}
\arguments{
\item{mu}{Array of mean matrices, possibly generated by function \code{generateMeans},
dimension should be \code{p}-by-\code{q}-by-\code{d}-by-\code{d}, where \code{p} is the number of matrices,
\code{q} is the number of SNR's to be used, and \code{d} is the dimension of square matrices.}

\item{cn}{Vector of convergence rattes in each replicate, set \code{n} as the sample size,
usually CLT indicates consistent estimators requires \code{n = cn^2}.}

\item{reps}{Number of replicates for the same distribution}

\item{nonneg}{Whether the samples should be all nonnegative.}
}
\value{
List of arrays.
\itemize{
\item mu.bar The array of consistent estimators of the mean matrices,
\item cov.bar The array of consistent estimators of covariance matrices.
}
}
\description{
The consistent estimators are obtained either as the mean of standard normally distributed matrices,
or as the estimated transition probability matrix of simulated Markov processes.
In addition the estimated covariance matrices are also returned.
}
\examples{
simuSamples(generateMeans(5,4,3), sqrt(c(50,100)), 50)
}
