% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hudsonDaily}
\alias{hudsonDaily}
\alias{hudsonWeekly}
\title{Streamflow Discharge Data}
\source{
\insertCite{usgs;textual}{eigTest}.
}
\usage{
hudsonDaily

hudsonWeekly
}
\description{
Hudson river discharges with quantiles and categories, consisting 1827 rows and 6 columns.
\itemize{
\item \code{hudsonDaily} - The discharge table of daily data.
\item \code{hudsonWeekly} - The discharge table of weekly data.
}
See \insertCite{xu2021testing;textual}{eigTest} for processing details.
}
\examples{
hudsonDaily; hudsonWeekly
}
\references{
\insertAllCited{}
}
\keyword{datasets}
