% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.test.R
\name{vec.test}
\alias{vec.test}
\title{Calculate the p-value for normal random vectors with specified covariance matrices}
\usage{
vec.test(Vlist, covList = list(), cm, testType = c("chi", "gam"))
}
\arguments{
\item{Vlist}{List of vectors}

\item{covList}{List of covariance matrices corresponding to the vectors}

\item{cm}{Constant for convergence}

\item{testType}{The test methods. Either for exact chi-squared test, or approximated gamma test}
}
\value{
A list of test information.
\itemize{
\item testType Test method, \code{'chi'} or \code{'gam'}.
\item statistic Test statistic.
\item df Degrees of freedom for chi-squared distribution.
\item shape Shape parameter in gamma distribution.
\item rate Rate parameter in gamma distribution.
\item pvalue P-value.
}
}
\description{
Calculate the p-value for normal random vectors with specified covariance matrices
}
\examples{
p = length(countryCoeff)
          vlist = vector('list', p)
          for (i in 1:p) {
            vlist[[i]] = as.double(countryCoeff[[i]])
          }
          vec.test(vlist, countryCovar, cm = sqrt(102), testType = 'chi')
}
