% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commutatorTest.R
\name{commutatorTest}
\alias{commutatorTest}
\title{Test on the commutator of two matrices}
\usage{
commutatorTest(
  matList,
  cm,
  d = nrow(A),
  covList = list(diag(d^2), diag(d^2)),
  testType = c("chi", "gam")
)
}
\arguments{
\item{matList}{The list of two matrices to be tested}

\item{cm}{Constant for convergence}

\item{d}{Size of matrices}

\item{covList}{The list of covariance matrices of the two matrices}

\item{testType}{The test methods. Either for exact chi-squared test, or approximated gamma test}
}
\value{
A list of test information.
\itemize{
\item testType Test method, \code{'chi'} or \code{'gam'}.
\item statistic Test statistic.
\item df Degrees of freedom for chi-squared distribution.
\item shape Shape parameter in gamma distribution.
\item rate Rate parameter in gamma distribution.
\item pvalue P-value.
}
}
\description{
Test on the commutator of two matrices
}
\examples{
means = generateMeans(5,2)
          samples = simuSamples(means, sqrt(100), 1)
          commutatorTest(samples[[1]][[1]][[1]][[1]], sqrt(400))
}
