% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commutatorTest.R
\name{commutatorTest}
\alias{commutatorTest}
\title{Test on the commutator of two matrices}
\usage{
commutatorTest(
  mat.arr,
  cn,
  cov.arr = NULL,
  testType = c("chi", "gam"),
  eps = NULL,
  param.out = FALSE
)
}
\arguments{
\item{mat.arr}{The array of two matrices to be tested with dimension \code{2}-\code{d}-\code{d}. If \code{dim(mat.arr)[1] > 2}, only the first two are used.}

\item{cn}{The convergence rate(s) to normality. Assume \code{n} is the sample size, usually CLT indicates \code{cn = sqrt(n)} for consistent estimators. If \code{length(cn) < p}, all matrices share the same rate \code{cn[1]}, otherwise \code{cn = cn[1:p]}.}

\item{cov.arr}{The array of covariance matrices of the two matrices with dimension \code{2}-\code{d^2}-\code{d^2}, default will use identity matrices when \code{is.null(cov.arr)}. If \code{dim(cov.arr)[1] > 2}, only the first two are used.}

\item{testType}{The test methods, can be exact chi-squared test \code{testType = 'chi'}, and/or approximated gamma test \code{testType = 'gam'}.}

\item{eps}{The threshold of eigenvalues when compute general inverse of covariance matrices. Required when \code{testType = 'chi'} but with default \code{cn^(-2/3)} when unsupplied.}

\item{param.out}{Logical, whether the parameters of limiting distribution should be output or not. Default \code{param.out = FALSE} to only output P-value.}
}
\value{
A named vector of P-value(s) when \code{param.out = FALSE}, or a named list of test information when \code{param.out = TRUE}, with name(s) to be \code{'chi'} and/or \code{'gam'}. Test information is a sub-list with elements including:
\itemize{
\item \code{testType}: The test methods. Either exact chi-squared test \code{'chi'}, or approximated gamma test \code{'gam'}.
\item \code{statistic}: The test statistic.
\item \code{df}: The degrees of freedom for chi-squared distribution when \code{testType = 'chi'}.
\item \code{shape}: The shape parameter in gamma distribution when \code{testType = 'gam'}.
\item \code{rate}: The rate parameter in gamma distribution when \code{testType = 'gam'}.
\item \code{pvalue}: The P-value.
}
}
\description{
Test on the commutator of two matrices
}
\examples{
means = generateMeans(5,2)
samples = simuSamples(means, sqrt(100), 1)
commutatorTest(samples[[1]][1,1,1,,,], sqrt(400))
}
