% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countryCoeff}
\alias{countryCoeff}
\title{Coefficient matrices of 8 countries' VAR(1) models.}
\format{
The array of coefficient matrices for country-wise macroeconomic VAR(1) models.
}
\usage{
countryCoeff
}
\description{
Coefficient matrices of 8 countries' VAR(1) models.
}
\examples{
\dontrun{
 countryCoeff
}
}
\keyword{datasets}
