% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countryCovar}
\alias{countryCovar}
\title{Asymptotic covariance matrices of the 8 coefficient matrices' estimates.}
\format{
The array of covariance matrices of the estimated VAR(1) coefficient matrices.
}
\usage{
countryCovar
}
\description{
Asymptotic covariance matrices of the 8 coefficient matrices' estimates.
}
\examples{
\dontrun{
 countryCovar
}
}
\keyword{datasets}
