% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countryMacro}
\alias{countryMacro}
\title{Macroeconomic Data}
\format{
The list of multivariate time series.
}
\source{
CEIC
}
\usage{
countryMacro
}
\description{
Standardized Macroeconomic Quarterly Data for 8 Countries.
Available from \insertCite{ceic,oecd;textual}{eigTest}.
See \insertCite{xu2021testing;textual}{eigTest}.
}
\examples{
\dontrun{
 countryMacro
}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
