% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countryMacro}
\alias{countryMacro}
\title{Standardized macroeconomic quarterly data for 8 countries}
\format{
Array of data frames
}
\source{
CEIC
}
\usage{
countryMacro
}
\description{
Standardized macroeconomic quarterly data for 8 countries
}
\examples{
\dontrun{
 countryMacro
}
}
\keyword{datasets}
