% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigTest.R
\name{eigTest}
\alias{eigTest}
\title{Test Common Eigenvectors for Multiple Matrices}
\usage{
eigTest(
  A,
  cn,
  cov.arr = NULL,
  V = NULL,
  testType = c("chi", "gam"),
  eps = NULL,
  param.out = FALSE
)
}
\arguments{
\item{A}{The array of matrices with dimension \code{p}-\code{d}-\code{d}, where \code{p} is the number of matrices, \code{d} is the dimension of the matrices.}

\item{cn}{The convergence rate(s) to normality. Assume \code{n} is the sample size, usually CLT indicates \code{cn = sqrt(n)} for consistent estimators. If \code{length(cn) < p}, all matrices share the same rate \code{cn[1]}, otherwise \code{cn = cn[1:p]}.}

\item{cov.arr}{The array of covariance matrices corresponding to the vectors with dimension \code{p}-\code{d^2}-\code{d^2}, default will use identity matrices when \code{is.null(cov.arr)}.}

\item{V}{The eigenvectors to be tested with dimension \code{d}-\code{d}. Default will use \code{V = JDTE(A)} when \code{is.null(V)} \insertCite{andre}{eigTest}.}

\item{testType}{The test methods, can be exact chi-squared test \code{testType = 'chi'}, and/or approximated gamma test \code{testType = 'gam'}.}

\item{eps}{The threshold of eigenvalues when compute general inverse of covariance matrices. Required when \code{testType = 'chi'} but with default \code{cn^(-2/3)} when unsupplied.}

\item{param.out}{Logical, whether the parameters of limiting distribution should be output or not. Default \code{param.out = FALSE} to only output p-value.}
}
\value{
A named vector of p-value(s) when \code{param.out = FALSE}, or a named list of test information when \code{param.out = TRUE}, with name(s) to be \code{'chi'} and/or \code{'gam'}. Test information is a sub-list with elements including:
\itemize{
\item \code{testType}: The test methods. Either exact chi-squared test \code{'chi'}, or approximated gamma test \code{'gam'}.
\item \code{statistic}: The test statistic.
\item \code{df}: The degrees of freedom for chi-squared distribution when \code{testType = 'chi'}.
\item \code{shape}: The shape parameter in gamma distribution when \code{testType = 'gam'}.
\item \code{rate}: The rate parameter in gamma distribution when \code{testType = 'gam'}.
\item \code{pvalue}: The p-value.
}
}
\description{
With approximated common eigenvector \code{V}, test whether the off-diagonal entries are annihilated after similarity transformations.
See \insertCite{xu2021testing;textual}{eigTest}.
}
\examples{
eigTest(countryCoeff, cn = sqrt(112), cov.arr = countryCovar, testType = 'gam')
}
\references{
\insertAllCited{}
}
