% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigTest.R
\name{eigTest}
\alias{eigTest}
\title{Test Whole Set of Eigenvectors}
\usage{
eigTest(
  A,
  cn,
  cov.arr = NULL,
  V = NULL,
  testType = c("chi", "gam"),
  eps = NULL,
  param.out = FALSE
)
}
\arguments{
\item{A}{Array of matrices}

\item{cn}{Constant for convergence}

\item{cov.arr}{List of covariance matrices, default with use identity matrices}

\item{V}{Eigenvector to be tested}

\item{testType}{The test methods. Either for exact chi-squared test, or approximated gamma test.}

\item{eps}{The threshold of eigenvalues when compute general inverse of covariance matrices. Required when \code{testType = 'chi'} but with default \code{cn^(-2/3)} when unsupplied}

\item{param.out}{The parameters of limiting distribution should be output or not}
}
\value{
A named vector of P-value(s) when \code{param.out = FALSE}, or a named list of test information when \code{param.out = TRUE}, with name(s) to be \code{'chi'} and/or \code{'gam'}. Test information is a sub-list with elements including:
\itemize{
\item statistic Test statistic.
\item df Degrees of freedom for chi-squared distribution.
\item shape Shape parameter in gamma distribution.
\item rate Rate parameter in gamma distribution.
\item pvalue P-value.
}
}
\description{
Test Whole Set of Eigenvectors
}
\examples{
eigTest(countryCoeff, cn = sqrt(112), cov.arr = countryCovar, testType = 'gam')
}
