% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expmPartSchur.R
\name{expmPartSchur}
\alias{expmPartSchur}
\title{Tangent space optimization for partially joint Schur decomposition using matrix exponential}
\usage{
expmPartSchur(A, k, warmup = FALSE, iter = 5000, tol = 10^(-12))
}
\arguments{
\item{A}{Array of matrices}

\item{k}{Number of Schur components}

\item{warmup}{Logical whether use \code{partSchur} for a warm-up \code{Q}}

\item{iter}{Maximum iteration number}

\item{tol}{Tolerance error}
}
\value{
Orthogonal matrix \code{Q}
}
\description{
Tangent space optimization for partially joint Schur decomposition using matrix exponential
}
\examples{
expmPartSchur(countryCoeff, 2)
}
