% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expmPartSchur.R
\name{expmPartSchur}
\alias{expmPartSchur}
\title{Tangent space optimization for partial Schur decomposition}
\usage{
expmPartSchur(
  A,
  k,
  d = dim(A)[2],
  p = dim(A)[1],
  iter = 5000,
  tol = 10^(-12),
  nn = FALSE
)
}
\arguments{
\item{A}{Array of matrices}

\item{k}{Number of Schur components}

\item{d}{Size of matrices}

\item{p}{Number of matrices}

\item{iter}{Maximum iteration number}

\item{tol}{Tolerance error}

\item{nn}{Logical whether the eigenvector elements are nonnegative}
}
\value{
Orthogonal matrix \code{Q}
}
\description{
Tangent space optimization for partial Schur decomposition
}
\examples{
expmPartSchur(countryCoeff, 2)
}
