% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMeans.R
\name{generateMeans}
\alias{generateMeans}
\title{Generate mean matrices for simulation samples}
\usage{
generateMeans(
  d,
  p,
  k = d,
  snr = 10,
  control.g = FALSE,
  V = NULL,
  v = NULL,
  nn = FALSE
)
}
\arguments{
\item{d}{The dimension of matrices.}

\item{p}{The number of matrices.}

\item{k}{The number of common Schur components. Must be an integer within (0, \code{d}), otherwise set \code{k = d}.}

\item{snr}{The positive signal to noise variance ratio (SNR), can be a vector.}

\item{control.g}{Logical, whether the control group of samples with perturbed common eigenvectors should be output or not.}

\item{V}{The input of eigenvector matrix with dimension \code{d}-\code{d}, needed when \code{nn = FALSE}. Default will use random sampling when \code{is.null(V)}.}

\item{v}{The input of stationary distribution with length \code{d}, needed when \code{nn = TRUE}. Default will use Dirichlet random sampling when \code{is.null(v)}.}

\item{nn}{Logical, whether the generated matrices should be nonnegative as transition probability matrices.}
}
\value{
An array of mean matrices: \code{p}-by-\code{q}-by-\code{d}-by-\code{d}, where \code{q} is the number of SNRs.
        If \code{control.g = TRUE}, \code{q = length(snr) + 1} otherwise \code{q = 1}.
}
\description{
Generate mean matrices for simulation samples
}
\examples{
generateMeans(5,8,3)
}
