% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hudsonDaily}
\alias{hudsonDaily}
\title{Streamflow Discharge Data}
\format{
The table of the Hudson river daily discharge with daily quantile and categories.
}
\usage{
hudsonDaily
}
\description{
Hudson river daily discharge table available from \insertCite{usgs}{eigTest}.
See \insertCite{xu2021testing;textual}{eigTest}.
}
\examples{
\dontrun{
 hudsonDaily
}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
