% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hudsonWeekly}
\alias{hudsonWeekly}
\title{Streamflow Discharge Data}
\format{
The table of the Hudson river weekly discharge with weekly quantile and categories.
}
\usage{
hudsonWeekly
}
\description{
Hudson river weekly discharge table available from \insertCite{usgs}{eigTest}.
See \insertCite{xu2021testing;textual}{eigTest}.
}
\examples{
\dontrun{
 hudsonWeekly
}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
