% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDTE.R
\name{JDTE}
\alias{JDTE}
\title{JDTE algorithm.}
\usage{
JDTE(A, d = dim(A)[2], p = dim(A)[1], iter = 5000, tol = 10^(-16))
}
\arguments{
\item{A}{Array of matrices.}

\item{d}{Size of matrices.}

\item{p}{Number of matrices.}

\item{iter}{Maximum iteration times.}

\item{tol}{Tolerance error.}
}
\value{
Matrix \code{U}.
}
\description{
JDTE algorithm.
}
\examples{
JDTE(countryCoeff)
}
