% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDTE.R
\name{JDTE}
\alias{JDTE}
\title{JDTE algorithm.}
\usage{
JDTE(A, iter = 500, tol = 10^(-6))
}
\arguments{
\item{A}{The array of matrices with dimension \code{p}-\code{d}-\code{d}, where \code{p} is the number of matrices, \code{d} is the dimension of the matrices.}

\item{iter}{The maximum iteration times.}

\item{tol}{The tolerance error for iteration termination.}
}
\value{
The eigenvector matrix \code{U} with dimension \code{d}-\code{d}.
}
\description{
JDTE algorithm.
}
\examples{
JDTE(countryCoeff)
}
