% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countryMacro}
\alias{countryMacro}
\alias{countryCoeff}
\alias{countryCovar}
\title{Macroeconomic Data}
\format{
`countryMacro` - The length-8 \code{list} of multivariate time series.

`countryCoeff` - The 8 x 3 x 3 \code{array} of coefficient matrices for country-wise macroeconomic VAR(1) models.

`countryCovar` - The 8 x 9 x 9 \code{array} of covariance matrices of the estimated VAR(1) coefficient matrices.
}
\source{
\insertCite{oecd,wbd,imf,FRB;textual}{eigTest} and other central bank statistics portals.\insertNoCite{CAN,GBR,DEU,FRA,JPN,KOR}{eigTest}
}
\usage{
countryMacro

countryCoeff

countryCovar
}
\description{
\itemize{
\item \code{countryMacro} - Macroeconomic Quarterly Data for 8 Countries.
\item \code{countryCoeff} - Estimated coefficient matrices of 8 countries' VAR(1) models.
\item \code{countryCovar} - Estimated covariance matrices of the 8 coefficient matrices' estimates.
}
See \insertCite{xu2021testing;textual}{eigTest} for processing details.
}
\examples{
countryMacro; countryCoeff; countryCovar
}
\references{
\insertAllCited{}
}
\keyword{datasets}
