% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countryCoeff}
\alias{countryCoeff}
\alias{countryCovar}
\title{Macroeconomic Data}
\format{
`countryCoeff` - The 8 x 3 x 3 \code{array} of coefficient matrices for country-wise macroeconomic VAR(1) models.

`countryCovar` - The 8 x 9 x 9 \code{array} of covariance matrices of the estimated VAR(1) coefficient matrices.
}
\source{
CEIC, OECD
}
\usage{
countryCoeff

countryCovar
}
\description{
\itemize{
\item \code{countryCoeff} - Estimated coefficient matrices of 8 countries' VAR(1) models.
\item \code{countryCovar} - Estimated covariance matrices of the 8 coefficient matrices' estimates.
}
Data are accessed from \insertCite{ceic,oecd;textual}{eigTest}.
See \insertCite{xu2021testing;textual}{eigTest} for processing details.
}
\examples{
countryCoeff; countryCovar
}
\references{
\insertAllCited{}
}
\keyword{datasets}
