% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnPartSchur.R
\name{nnPartSchur}
\alias{nnPartSchur}
\title{Joint Schur decomposition for the non-negative leading component}
\usage{
nnPartSchur(A)
}
\arguments{
\item{A}{Array of matrices}
}
\value{
Orthogonal matrix \code{Q}
}
\description{
Joint Schur decomposition for the non-negative leading component
}
\examples{
nnPartSchur(countryCoeff)
}
