% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnPartSchur.R
\name{nnPartSchur}
\alias{nnPartSchur}
\title{Optimization of the Common Non-negative Leading Eigenvector}
\usage{
nnPartSchur(A)
}
\arguments{
\item{A}{The array of matrices with dimension \code{p}-\code{d}-\code{d}, where \code{p} is the number of matrices, \code{d} is the dimension of the matrices.}
}
\value{
The orthogonal matrix \code{Q} of dimension \code{d}-\code{d} with the first column to be the estimated nonnegative common component.
}
\description{
Joint Schur decomposition, usually applicable on non-negative square matrices such as transition probability matrices, to estimate the common leading eigenvector.
See \insertCite{xu2021testing;textual}{eigTest}.
}
\examples{
nnPartSchur(countryCoeff)
}
\references{
\insertAllCited{}
}
