% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialTest.R
\name{partialTest}
\alias{partialTest}
\title{Test Subset of Common Eigenvectors}
\usage{
partialTest(
  A,
  cn,
  cov.arr = NULL,
  nn = FALSE,
  k = NULL,
  warmup = FALSE,
  Q = NULL,
  V = NULL,
  testType = c("chi", "gam"),
  eps = NULL,
  param.out = FALSE
)
}
\arguments{
\item{A}{The array of matrices with dimension \code{p}-\code{d}-\code{d}, where \code{p} is the number of matrices, \code{d} is the dimension of the matrices.}

\item{cn}{The convergence rate(s) to normality. Assume \code{n} is the sample size, usually CLT indicates \code{cn = sqrt(n)} for consistent estimators. If \code{length(cn) < p}, all matrices share the same rate \code{cn[1]}, otherwise \code{cn = cn[1:p]}.}

\item{cov.arr}{The array of covariance matrices corresponding to the vectors with dimension \code{p}-\code{d^2}-\code{d^2}, default will use identity matrices when \code{is.null(cov.arr)}.}

\item{nn}{Logical, whether the matrices \code{A} are regarded as nonnegative transition probability matrices and the the eigenvector elements should be nonnegative.}

\item{k}{The number of common components to be tested. When \code{nn=TRUE}, \code{k} can be \code{NULL} or \code{k = 1}, otherwise \code{k} must be an integer within (0, \code{d}]. Will call \code{eigTest} instead if \code{k = d}.}

\item{warmup}{Logical, whether use \code{partSchur} for a warm-up initial value, default to \code{FALSE}.}

\item{Q}{The orthogonal components to be tested with dimension \code{d}-\code{d}. Default (when \code{is.null(Q)}) will use \code{Q = nnPartSchur(A)} if \code{nn = TRUE} else \code{Q = expmPartSchur(A, k, warmup = warmup)}.}

\item{V}{The \code{k}-\code{k} eigenvector matrix to be tested for the upper diagonal block after orthogonally transformed by \code{Q}. Only applicable when \code{nn = FALSE}. Default (when \code{is.null(V)}) will use \code{V = JDTE(A)} if \code{k = d} else \code{V = JDTE(B)} where \code{B[i,,] = (t(Q) A[i,,] Q)[1:k,1:k]}.}

\item{testType}{The test methods, can be exact chi-squared test \code{testType = 'chi'}, and/or approximated gamma test \code{testType = 'gam'}.}

\item{eps}{The threshold of eigenvalues when compute general inverse of covariance matrices. Required when \code{testType = 'chi'} but with default \code{cn^(-2/3)} when unsupplied.}

\item{param.out}{Logical, whether the parameters of limiting distribution should be output or not. Default \code{param.out = FALSE} to only output P-value.}
}
\value{
A named vector of P-value(s) when \code{param.out = FALSE}, or a named list of test information when \code{param.out = TRUE}, with name(s) to be \code{'chi'} and/or \code{'gam'}. Test information is a sub-list with elements including:
\itemize{
\item \code{testType}: The test methods. Either exact chi-squared test \code{'chi'}, or approximated gamma test \code{'gam'}.
\item \code{statistic}: The test statistic.
\item \code{df}: The degrees of freedom for chi-squared distribution when \code{testType = 'chi'}.
\item \code{shape}: The shape parameter in gamma distribution when \code{testType = 'gam'}.
\item \code{rate}: The rate parameter in gamma distribution when \code{testType = 'gam'}.
\item \code{pvalue}: The P-value.
}
}
\description{
Test whether a subset of common eigenvectors are shared by the list of random matrices.
}
\examples{
partialTest(countryCoeff, cn = sqrt(112), cov.arr = countryCovar, k = 2, testType = 'gam')
}
\references{
\insertRef{xu2021testing}{eigTest}
    \insertRef{Flury86}{eigTest}
    \insertRef{tensor}{eigTest}
    \insertRef{andre}{eigTest}
}
