% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialTest.R
\name{partialTest}
\alias{partialTest}
\title{Test Subset of Common Eigenvectors}
\usage{
partialTest(
  A,
  covList = list(),
  k,
  cn,
  nn = FALSE,
  Q = expmPartSchur(A, k, nn = nn),
  n = nrow(A[[1]]),
  p = length(A),
  testType = c("chi", "gam"),
  param.out = FALSE
)
}
\arguments{
\item{A}{List of original matrices}

\item{covList}{List of original covariance matrices, default will use identity matrices}

\item{k}{Size of the block matrices to be tested}

\item{cn}{Constant for convergence}

\item{nn}{Logical whether the eigenvector elements are nonnegative}

\item{Q}{Schur components for transformation}

\item{n}{Size of original matrices}

\item{p}{Number of matrices}

\item{testType}{The test methods. Either for exact chi-squared test, or approximated gamma test}

\item{param.out}{The parameters of limiting distribution should be output or not}
}
\value{
P-value or a list of test information.
\itemize{
\item testType Test method, \code{'chi'} or \code{'gam'}.
\item statistic Test statistic.
\item df Degrees of freedom for chi-squared distribution.
\item shape Shape parameter in gamma distribution.
\item rate Rate parameter in gamma distribution.
\item pvalue P-value.
}
}
\description{
Test whether a subset of common eigenvectors are shared by the list of random matrices.
}
\examples{
partialTest(countryCoeff, countryCovar, k = 2, cn = 102, testType = 'gam')
}
