% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partSchur.R
\name{partSchur}
\alias{partSchur}
\title{Partial Joint Schur Decomposition}
\usage{
partSchur(A, k, iter = 5000, tol = 10^(-16))
}
\arguments{
\item{A}{The array of matrices with dimension \code{p}-\code{d}-\code{d}, where \code{p} is the number of matrices, \code{d} is the dimension of the matrices.}

\item{k}{The number of common components to be tested. \code{k} must be an integer within (0, \code{d}).}

\item{iter}{The maximum iteration number.}

\item{tol}{The tolerance error for iteration termination.}
}
\value{
The orthogonal matrix \code{Q} of dimension \code{d}-\code{d} with the first \code{k} columns to be the estimated common components.
}
\description{
Partial Joint Schur Decomposition
}
\keyword{internal}
