% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partSchur.R
\name{partSchur}
\alias{partSchur}
\title{Partial Joint Schur Decomposition}
\usage{
partSchur(A, k, iter = 5000, tol = 10^(-16))
}
\arguments{
\item{A}{Array of matrices}

\item{k}{Number of Schur components}

\item{iter}{Maximum iteration number}

\item{tol}{Tolerance error}

\item{nonneg}{Logical whether the eigenvector elements are nonnegative}
}
\value{
Orthogonal matrix \code{Q}
}
\description{
Partial Joint Schur Decomposition
}
\keyword{internal}
