% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projTest.R
\name{projTest}
\alias{projTest}
\title{Test on the Log-Likelihood Ratio (LLR) for Two Matrices}
\usage{
projTest(
  A,
  cn,
  cov.arr = NULL,
  eps = NULL,
  refMat = NULL,
  poly.sp = TRUE,
  CV = NULL,
  param.out = FALSE
)
}
\arguments{
\item{A}{The array of two matrices to be tested with dimension \code{2}-\code{d}-\code{d}. If \code{dim(A)[1] > 2}, only the first two are used.}

\item{cn}{The convergence rate to normality. Assume \code{n} is the sample size, usually CLT indicates \code{cn = sqrt(n)} for consistent estimators.}

\item{cov.arr}{The array of covariance matrices of the two matrices with dimension \code{2}-\code{d^2}-\code{d^2}, default will use identity matrices when \code{is.null(cov.arr)}. If \code{dim(cov.arr)[1] > 2}, only the first two are used.}

\item{eps}{The threshold of eigenvalues when compute general inverse of covariance matrices. Required when \code{testType = 'chi'} but with default \code{cn^(-2/3)} when unsupplied.}

\item{refMat}{The array of reference matrices with dimension \code{2}-\code{d}-\code{d} that have the same eigenvectors. Default (when \code{is.null(refMat)}) is to use \code{A[2,,]} as reference for \code{A[1,,]} and vice versa. If \code{dim(refMat)[1] = 1}, the only reference matrix is shared.}

\item{poly.sp}{Logical, whether the space matrix is generated from polynomial basis. Default \code{poly.sp = TRUE} to use Legendre polynomials. Otherwise \code{poly.sp = FALSE}, the space matrix is generated with common eigenvectors.}

\item{CV}{Matrix of dimension \code{d}-by-\code{d} only functionable when \code{poly.sp = FALSE}, as the supplied reference common eigenvector matrix. Default (when \code{is.null(V)}) is to call \code{V = JDTE(refMat)} \insertCite{andre}{eigTest}.}

\item{param.out}{Logical, whether the parameters of limiting distribution should be output or not. Default \code{param.out = FALSE} to only output p-value.}
}
\value{
A p-value when \code{param.out = FALSE} or a list of test information when \code{param.out = TRUE}.
\itemize{
\item \code{statistic}: The test statistic.
\item \code{df}: The degrees of freedom for chi-squared distribution.
\item \code{pvalue} The p-value.
}
}
\description{
Only valid for nonsingular covariance matrices \code{cov.arr}.
}
\details{
For the LLR test, the maximum likelihood estimators (MLEs) are obtained by covariance-weighted projections onto the linear space either characterized by matrix polynomials or approximated common eigenvectors.
See \insertCite{xu2021testing;textual}{eigTest}.
}
\examples{
projTest(countryCoeff, cn = sqrt(112), countryCovar)
}
\references{
\insertAllCited{}
}
