% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projTest.R
\name{projTest}
\alias{projTest}
\title{Calculate the p-value for projected MLE with specified covariance matrices}
\usage{
projTest(
  A,
  cov.arr = NULL,
  refMat = A,
  cn,
  eps,
  d = dim(A)[2],
  param.out = FALSE
)
}
\arguments{
\item{A}{Array of matrices to be tested. Require the length to be 2.}

\item{cov.arr}{List of covariance matrices corresponding to the random matrices, default will use identity matrices}

\item{refMat}{The list of reference matrices with the same eigenvectors, default is to use the estimated \code{A}}

\item{cn}{Constant for convergence}

\item{eps}{The threshold of eigenvalues when compute general inverse of covariance matrices}

\item{d}{Matrix dimension}

\item{param.out}{Logical. Whether the parameters need to be included in output}
}
\value{
The p-value or a list of test information.
\itemize{
\item statistic Test statistic.
\item df Degrees of freedom for chi-squared distribution.
\item pvalue P-value.
}
}
\description{
Only valid for nonsingular covariance matrices \code{cov.arr}.
}
\examples{
projTest(countryCoeff, countryCovar, cn = sqrt(112), eps = 112^(-1/3))
}
