% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projTest.R
\name{projTest}
\alias{projTest}
\title{Calculate the p-value for projected MLE with specified covariance matrices}
\usage{
projTest(A, covList = list(), cn, n = ncol(A[[1]]), param.out = FALSE)
}
\arguments{
\item{A}{List of matrices to be tested. Require the length to be 2.}

\item{covList}{List of covariance matrices corresponding to the random matrices, default will use identity matrices}

\item{cn}{Constant for convergence}

\item{n}{Matrix dimension}

\item{param.out}{Logical. Whether the parameters need to be included in output}
}
\value{
The p-value or a list of test information.
\itemize{
\item statistic Test statistic.
\item df Degrees of freedom for chi-squared distribution.
\item pvalue P-value.
}
}
\description{
Calculate the p-value for projected MLE with specified covariance matrices
}
\examples{
projTest(countryCoeff, countryCovar, cn = 102)
}
