% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schurTest.R
\name{schurTest}
\alias{schurTest}
\title{Test Schur Components}
\usage{
schurTest(
  A,
  covList = list(),
  k,
  cn,
  nn = FALSE,
  Q = NULL,
  n = ncol(A[[1]]),
  p = length(A),
  testType = c("gam", "chi"),
  param.out = FALSE
)
}
\arguments{
\item{A}{List of matrices}

\item{covList}{List of covariance matrices, default will use identity matrices}

\item{k}{Number of components to be tested}

\item{cn}{Constant for convergence}

\item{nn}{Logical whether the eigenvector elements are nonnegative}

\item{Q}{Orthogonal components to be tested}

\item{n}{Size of matrices}

\item{p}{Number of matrices}

\item{testType}{The test methods. Either for exact chi-squared test, or approximated gamma test}

\item{param.out}{The parameters of limiting distribution should be output or not}
}
\value{
P-value or a list of test information.
\itemize{
\item testType Test method, \code{'chi'} or \code{'gam'}.
\item statistic Test statistic.
\item df Degrees of freedom for chi-squared distribution.
\item shape Shape parameter in gamma distribution.
\item rate Rate parameter in gamma distribution.
\item pvalue P-value.
}
}
\description{
Test Schur Components
}
\examples{
schurTest(countryCoeff, countryCovar, k = 2, cn = 102, testType = 'chi')
}
