% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.fun.R
\name{score.fun}
\alias{score.fun}
\title{Score function (first semi-column squared norm)}
\usage{
score.fun(A, Q)
}
\arguments{
\item{A}{List of matrices}

\item{Q}{The orthogonal matrix}
}
\value{
The sum of squared norms
}
\description{
Calculate the summation of squared 'F' norms of \code{Bi[(j+1):n,j]} where \code{Bi = t(Q) Ai Q}.
}
\keyword{internal}
