% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.fun.R
\name{score.fun}
\alias{score.fun}
\title{Score function (first semi-column squared norm)}
\usage{
score.fun(A, Q)
}
\arguments{
\item{A}{The array of matrices with dimension \code{p}-\code{d}-\code{d}, where \code{p} is the number of matrices, \code{d} is the dimension of the matrices.}

\item{Q}{The orthogonal matrix of dimension \code{d}-\code{d} with the first column evaluated for common.}
}
\value{
The sum of squared norms of \code{t(Q[,-1]) Ai Q[,1]}.
}
\description{
Calculate the summation of squared 'F' norms of \code{Bi[(j+1):n,j]} where \code{Bi = t(Q) Ai  Q}.
}
\keyword{internal}
