% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuSamples.R
\name{simuSamples}
\alias{simuSamples}
\title{Generating consistent estimators for given means}
\usage{
simuSamples(mu, cn, samples, nonneg = FALSE)
}
\arguments{
\item{mu}{Array of mean matrices, possibly generated by function \code{generateMeans}}

\item{cn}{Vector of constants for convergence in each replicate, usually \code{sqrt(n)} in CLT,
i.e. the sample size to get the consistent estimators is \code{n = M^2}.}

\item{samples}{Number of replicates for the same distribution}

\item{nonneg}{Whether the samples should be all nonnegative.}
}
\value{
List of lists.
\itemize{
\item mu.bar The list of consistent estimators of the mean matrices,
\item cov.bar The list of consistent estimators of covariance matrices.
}
}
\description{
The consistent estimators are obtained either as the mean of standard normally distributed matrices,
or as the estimated transition probability matrix of simulated Markov processes.
In addition the estimated covariance matrices are also returned.
}
\examples{
simuSamples(generateMeans(5,8,3), sqrt(c(100,500)), 100)
}
