% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuSamples.R
\name{simuSamples}
\alias{simuSamples}
\title{Generating consistent estimators for given means}
\usage{
simuSamples(mu, cn, reps, nn = FALSE, est.cov = TRUE, prl = FALSE)
}
\arguments{
\item{mu}{The array of mean matrices, possibly generated by function \code{generateMeans}. Dimension should be \code{p}-by-\code{q}-by-\code{d}-by-\code{d}, where \code{p} is the number of matrices, \code{q} is the number of SNR's to be used, and \code{d} is the dimension of square matrices.}

\item{cn}{The vector of convergence rates in each replicate. Assume \code{n} is the sample size, usually CLT indicates \code{cn = sqrt(n)} for consistent estimators.}

\item{reps}{The number of replicates for the same distribution.}

\item{nn}{Logical, whether the generated matrices should be nonnegative as transition probability matrices. Default to \code{nn = FALSE}}

\item{est.cov}{Logical, whether the covariance matrices are estimated empirically. Default to \code{est.cov = TRUE}.}

\item{prl}{Logical, whether to use parallel sampling. Default to \code{prl = FALSE} for not parallelizing. If \code{prl == TRUE}, one must first call \code{registerDoSNOW()} externally.}
}
\value{
A list of lists, with length \code{reps * m * q}, where \code{m} is the number of different rates \code{m = length(cn)}. Every sublist has the following format.
\itemize{
\item \code{SNR}: The string indicating the signal to noise ratio.
\item \code{CovRate}: The numerical value indicating the corresponding convergence rate in \code{cn}.
\item \code{mu.bar}: The array of consistent estimators of the mean matrices, with dimension \code{p}-\code{d}-by-\code{d}.
\item \code{cov.bar}: The array of consistent estimators of covariance matrices, with dimension \code{p}-\code{d^2}-by-\code{d^2}.
}
When \code{est.cov = FALSE}, only the first array of mean estimators are included in the sublist.
}
\description{
The consistent estimators are obtained either as the mean of standard normally distributed matrices,
or as the estimated transition probability matrix of simulated Markov processes.
In addition the estimated covariance matrices are also returned.
}
\examples{
simuSamples(generateMeans(5,4,3), sqrt(c(50,100)), 50)
}
