% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hudsonDaily}
\alias{hudsonDaily}
\alias{hudsonWeekly}
\title{Streamflow Discharge Data}
\usage{
hudsonDaily

hudsonWeekly
}
\description{
Hudson river discharges with quantiles and categories, consisting 1827 rows and 6 columns.
\itemize{
\item `hudsonDaily` - The discharge table of daily data.
\item `hudsonWeekly` - The discharge table of weekly data.
}
Data are accessed from \insertCite{usgs;textual}{eigTest}.
See \insertCite{xu2021testing;textual}{eigTest} for processing details.
}
\examples{
hudsonDaily; hudsonWeekly
}
\references{
\insertAllCited{}
}
\keyword{datasets}
