% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateSVD.R
\name{truncateSVD}
\alias{truncateSVD}
\title{Compute truncated SVD of a matrix.}
\usage{
truncateSVD(A, eps)
}
\arguments{
\item{A}{Matrix}

\item{eps}{The truncation threshold of singular values for SVD}
}
\value{
A list of information about truncated SVD.
\itemize{
\item Id The truncated approximation of \code{A}.
\item r The rank of \code{A} after truncation.
\item ginv The general inverse of the truncated SVD approximation.
}
}
\description{
Compute truncated SVD of a matrix.
}
\keyword{internal}
