% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateSVD.R
\name{truncateSVD}
\alias{truncateSVD}
\title{Compute truncated SVD of a (positive semi-definite) matrix.}
\usage{
truncateSVD(A, eps, tr.approx = FALSE)
}
\arguments{
\item{A}{The square matrix of dimension \code{d}-\code{d} to be decomposed.}

\item{eps}{The truncation threshold of singular values for SVD.}

\item{tr.approx}{Logical, whether approximate the truncated version of \code{A}}
}
\value{
A list of information about truncated SVD.
\itemize{
\item \code{Id}: The truncated approximation of \code{A}. Included only when \code{tr.approx = TRUE}.
\item \code{r}: The rank of \code{A} after truncation.
\item \code{rootdinv.u} The left square root matrix of \code{A}, i.e., the truncated generalized inverse \eqn{A^+ \approx}\code{tcrossprod(rootdinv.u)}.
}
}
\description{
Compute truncated SVD of a (positive semi-definite) matrix.
}
\keyword{internal}
