% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateList.R
\name{updateList}
\alias{updateList}
\title{Update the list of matrices}
\usage{
updateList(A, Q)
}
\arguments{
\item{A}{Original array of matrices}

\item{Q}{Orthogonal matrix}
}
\value{
Array of transformed matrices
}
\description{
Calculate the list of matrices after orthogonal transformation by matrix \code{Q}
}
\keyword{internal}
