% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec.test.R
\name{vec.test}
\alias{vec.test}
\title{Calculate the p-value for normal random vectors with specified covariance matrices}
\usage{
vec.test(V.arr, cn, testType, cov.arr = NULL, eps = NULL, param.out = FALSE)
}
\arguments{
\item{V.arr}{Array of vectors}

\item{cn}{The constant for convergence, or a vector of constants.}

\item{testType}{The test methods. Either for exact chi-squared test, or approximated gamma test}

\item{cov.arr}{Array of covariance matrices corresponding to the vectors, default will use identity matrices}

\item{eps}{The threshold of eigenvalues when compute general inverse of covariance matrices. Required when \code{testType = 'chi'} but with default \code{cn^(-2/3)} when unsupplied}

\item{param.out}{The parameters of limiting distribution should be output or not}
}
\value{
A list of test information.
\itemize{
\item statistic Test statistic.
\item df Degrees of freedom for chi-squared distribution.
\item shape Shape parameter in gamma distribution.
\item rate Rate parameter in gamma distribution.
\item pvalue P-value.
}
}
\description{
Calculate the p-value for normal random vectors with specified covariance matrices
}
\keyword{internal}
